# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Phonebook API
  version: 1.0.0
  description: A reactive REST API for managing phonebook contacts

servers:
  - url: http://localhost:8000

paths:
  /api/contacts:
    get:
      summary: Get a paginated list of contacts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_Contact'
    post:
      summary: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: Contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/contacts/search:
    get:
      summary: Search contacts by name or phone
      parameters:
        - in: query
          name: query
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_Contact'

  /api/contacts/{id}:
    get:
      summary: Get a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contact found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'

    delete:
      summary: Delete a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Contact deleted

components:
  schemas:
    Contact:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        formattedPhoneNumber:
          type: string
        address:
          type: string
      required: [firstName, phoneNumber, formattedPhoneNumber]

    PagedResponse_Contact:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        totalElements:
          type: integer
        page:
          type: integer
        size:
          type: integer

    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        requestPath:
          type: string